+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.0
+
+* Support adventure scene adjustments exported just for v10. Some scene data would not be altered correctly in v10 if the scene data had been exported in v10.
+
 # 1.1.2
 
 * Note pins would be generated for some adventures, in some circumstances, even if you didn't want them to be.